+++ CHANGELOG.md
@@ -9,11 +9,15 @@ All notable changes to this project will be documented in this file.
 - The newly added `<Editor />` replaces the removed `<EditorProvider />`. It sets up the `EditorContext` and renders an editor for the given `state`.
 - Don't export API that isn't intended for external usage anymore.
 - Rename `createDocumentIdentifier` to `createDocument`
+- Remove `StatefulPlugin.createInitialState` in favor of `StatefulPlugin.state`
+- Plugins no longer receive `onChange`. Use the provided setters & helpers on `state` instead
 
 ### Added
 
 - Handle serialization and deserialization of documents
 - Add `<Editor />` that replaces `<EditorProvider />`
+- Add `StateType.scalar` that represents a value of the given type (more specifically: `StateType.boolean`, `StateType.number`, `StateType.child`
+- Add `StateType.list`, `State.object`
 
 ### Changed